% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{BindBuildCurrentPostForJekyllAsDraft}
\alias{BindBuildCurrentPostForJekyllAsDraft}
\title{Build the current file as draft}
\usage{
BindBuildCurrentPostForJekyllAsDraft()
}
\description{
Build the current file as draft
}

