% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{NewLabnotebook}
\alias{NewLabnotebook}
\title{Create a file in labnotebook.dir then open it and build it with knitr and jekyll.}
\usage{
NewLabnotebook(title, labnotebook.dir = getOption("Mokusu.labnotebook.dir"),
  jekyll.dir = getOption("Mokusu.jekyll.dir"),
  skeleton.file = ".skeleton_post")
}
\description{
Create a file in labnotebook.dir then open it and build it with knitr and jekyll.
}

