% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{ServeJekyllWebsite}
\alias{ServeJekyllWebsite}
\title{Serve Jekyll}
\usage{

  ServeJekyllWebsite(jekyll.dir = "/home/cayek/PatatorHomeDir/Projects/PersonalWebSite/",
  draft = TRUE)
}
\description{
Serve Jekyll
}

