% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packratEnv.R
\name{UpdateGithubPkg}
\alias{UpdateGithubPkg}
\title{Update packrate env}
\usage{
UpdateGithubPkg(pkgs = c("Mokusu", "associationr", "BioCompToolsR", "tess3r",
  "tess3rOld"), checkonly = TRUE)
}
\description{
Update packrate env
}

