% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packratEnv.R
\name{UpdateGithubPkgPackrat}
\alias{UpdateGithubPkgPackrat}
\title{Update packrate env}
\usage{
UpdateGithubPkgPackrat(pkgs = c("Mokusu", "TESS3enchoSen", "associationr",
  "BioCompToolsR", "tess3r"))
}
\description{
Update packrate env
}

