% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blog.R
\name{NewLabnotebook}
\alias{NewLabnotebook}
\title{Create a file in labnotebook.dir then open it and build it with knitr and jekyll.}
\usage{
NewLabnotebook(title,
  labnotebook.dir = "/home/cayek/PatatorHomeDir/Notebook/Labnotebook/",
  jekyll.dir = "~/PatatorHomeDir/Projects/PersonalWebSite/",
  skeleton.file = ".skeleton_post")
}
\description{
Create a file in labnotebook.dir then open it and build it with knitr and jekyll.
}

